sabekov/requirements.txt
django
django-pipeline
django-widget-tweaks
django-crispy-forms
